// src/@types/actors/general/aids/aidsResponse.type.ts
import { TYPE_AIDS } from "@/content/actor/manager/aids-management";

export interface general_Aid {
    id: number | string;
    type: TYPE_AIDS;
    distribution_date: string;
    recipients_number: number;
    title?: string;
    description?: string;
    complete: boolean
}

export interface AidsResponse {
    status: string;
    message?: string;
    aids: general_Aid[];
    pagination: {
        page: number;
        limit: number;
        totalItems: number;
        totalPages: number;
    };
    error?: string;
}